Move standard issue filter to dop.
