+++ internal/apps/dop/component-protocol/components/project-pipeline/myPage/tabsTable/filterContainer/customFilter/condition.go
@@ -20,8 +20,8 @@ import (
 	"github.com/erda-project/erda/apistructs"
 	"github.com/erda-project/erda/internal/apps/dop/component-protocol/components/project-pipeline/common"
 	"github.com/erda-project/erda/internal/apps/dop/component-protocol/components/util"
+	"github.com/erda-project/erda/internal/apps/dop/component-protocol/standard-components/condition"
 	"github.com/erda-project/erda/internal/apps/dop/providers/projectpipeline/deftype"
-	"github.com/erda-project/erda/internal/pkg/component-protocol/condition"
 	"github.com/erda-project/erda/pkg/limit_sync_group"
 )
