+++ internal/apps/dop/component-protocol/standard-components/condition/input.go
